+++ source/extensions/omni.flux.asset_importer.core/docs/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.16.10]
+### Fixed
+- Fixing scan folder dialog issues
+
 ## [1.16.9]
 ### Added
 - Add a scan folder function